[refactor]:move all exception classes to ecnu.db.exception
