+++ src/main/java/ecnu/db/analyzer/statical/QueryTableName.java
@@ -5,7 +5,7 @@ import com.alibaba.druid.sql.ast.SQLStatement;
 import com.alibaba.druid.sql.visitor.SchemaStatVisitor;
 import com.alibaba.druid.stat.TableStat;
 import ecnu.db.utils.CommonUtils;
-import ecnu.db.utils.TouchstoneToolChainException;
+import ecnu.db.exception.TouchstoneToolChainException;
 
 import java.util.HashSet;
 import java.util.List;
