+++ src/main/java/ecnu/db/schema/AbstractSchemaGenerator.java
@@ -4,7 +4,7 @@ import ecnu.db.dbconnector.AbstractDbConnector;
 import ecnu.db.schema.Schema;
 import ecnu.db.schema.column.*;
 import ecnu.db.utils.ConfigConvert;
-import ecnu.db.utils.TouchstoneToolChainException;
+import ecnu.db.exception.TouchstoneToolChainException;
 import org.apache.commons.lang3.tuple.Pair;
 
 import java.io.IOException;
