+++ src/main/java/ecnu/db/tidb/TidbNodeTypeTool.java
