+++ src/main/java/ecnu/db/tidb/stats/Histogram.java
@@ -1,0 +1,2 @@
+package ecnu.db.tidb.stats;public class Histogram {
+}
