+++ src/main/java/ecnu/db/utils/SqlTemplateHelper.java
@@ -7,7 +7,7 @@ import com.alibaba.druid.util.JdbcConstants;
 import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.Multimap;
 import ecnu.db.constraintchain.filter.Parameter;
-import ecnu.db.utils.exception.UnsupportedDBTypeException;
+import ecnu.db.exception.UnsupportedDBTypeException;
 import org.apache.commons.lang3.tuple.Pair;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
